Done Ekinomy Basic Economic Manager

+++ src/main/kotlin/com/chaos/ekinomy/data/PlayerBalanceData.kt
@@ -2,12 +2,12 @@ package com.chaos.ekinomy.data
 
 import java.util.*
 
-data class PlayerBalanceData(val playerName: String, val playerUUID: UUID, var balance: Double) {
-    operator fun plus(balance: Double) = PlayerBalanceData(playerName, playerUUID, this.balance + balance)
+data class PlayerBalanceData(val playerName: String, val playerUUID: UUID, var balance: Long) {
+    operator fun plus(balance: Long) = PlayerBalanceData(playerName, playerUUID, this.balance + balance)
 
     operator fun plus(balanceData: PlayerBalanceData) = plus(balanceData.balance)
 
-    operator fun plusAssign(balance: Double) {
+    operator fun plusAssign(balance: Long) {
         this.balance += balance
     }
